pyproject.toml
[project]
name = "stemset"
version = "0.1.0"
description = "AI-powered audio stem separation with web playback interface"
requires-python = "==3.13.*"
dependencies = [
    # API/Web server (lightweight - deployed to Koyeb)
    "litestar>=2.0.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.1.1",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.9",
    "modal>=0.64.0",
]

[project.scripts]
stemset = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "black>=23.0.0", "ruff>=0.1.0"]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = ["basedpyright>=1.31.7"]
preload_models = ["audio-separator>=0.39.1"]
shared = [
    # Shared dependencies used by both API and processing
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "boto3>=1.28.0,<1.33.0",
    "typer>=0.20.0",
]
processing = [
    # Heavy ML/audio processing dependencies (local/gpu-worker)
    "audio-separator>=0.39.1",
    "pyloudnorm>=0.1.1",
    "soundfile>=0.12.1",
    "watchdog>=3.0.0",
    "opuslib>=3.0.1",
    "onnxruntime>=1.23.1",
    "numpy>=2.3.4",
    "pillow>=10.0.0",
]
modal = [
    # Dependencies needed for Modal deployment
    "fastapi>=0.115.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.uv]
allow-insecure-host = ["*.r2.cloudflarestorage.com"]
default-groups = ["shared", "processing"]
